combine type files
