+++ src/shared/relay/index.js
@@ -1,5 +1,6 @@
 // @flow
 export { currentEnvironment } from './currentEnvironment'
+export { unwrap } from './unwrap'
 export { withLoadMoreArgs } from './withLoadMoreArgs'
 export { cacheResolvers } from './cacheResolvers'
 export { createPaginationCacheResolver } from './createPaginationCacheResolver'
