+++ src/shared/relay/unwrap.js
@@ -1,0 +1,6 @@
+// @flow
+import type { Connection } from 'shared/types'
+
+export function unwrap<T> (connection: ?Connection<T>): Array<T> {
+  return connection ? connection.edges.map((edge) => edge.node) : []
+}
