+++ src/shared/styles/index.js
@@ -6,3 +6,4 @@ export { borders } from './borders'
 export { stylesheet } from './stylesheet'
 export { mobile } from './mobile'
 export { utils } from './utils'
+export { combine } from './combine'
