--- src/shared/types/Chamber.js
@@ -1,4 +1,0 @@
-// @flow
-export type Chamber = {
-  name: string,
-}
