--- src/shared/types/Committee.js
@@ -1,4 +1,0 @@
-// @flow
-export type Committee = {
-  name: string,
-}
