--- src/shared/types/Edge.js
@@ -1,5 +1,0 @@
-// @flow
-export type Edge<T> = {
-  node: T,
-  cursor: string
-}
