--- src/shared/types/Hearing.js
@@ -1,5 +1,0 @@
-// @flow
-export type Hearing = {
-  id: string,
-  date: string,
-}
