--- src/shared/types/History.js
@@ -1,4 +1,0 @@
-// @flow
-export type History = {
-  goBack (): void
-}
