--- src/shared/types/Location.js
@@ -1,4 +1,0 @@
-// @flow
-export type Location = {
-  action: 'PUSH' | 'POP'
-}
