--- src/shared/types/PageInfo.js
@@ -1,4 +1,0 @@
-// @flow
-export type PageInfo = {
-  hasNextPage: boolean
-}
