--- src/shared/types/RelayProp.js
@@ -1,9 +1,0 @@
-// @flow
-type Variables = {[key: string]: any}
-
-export type RelayProp = {
-  variables: Variables,
-  pendingVariables: ?Variables,
-  setVariables (partialVariables: Object, onReadyStateChange?: Function): void,
-  forceFetch (partialVariables: Object, onReadyStateChange?: Function): void
-}
