--- src/shared/types/Viewer.js
@@ -1,9 +1,0 @@
-// @flow
-import type { SearchConnection } from './Connection'
-import type { Bill } from './Bill'
-
-export type Viewer = {
-  isAdmin: boolean,
-  bill: Bill,
-  bills: SearchConnection<Bill>
-}
