+++ src/shared/types/client.js
@@ -1,4 +1,6 @@
 // @flow
+export type StyleProp = Object | Array<Object>
+
 export type Notification = {
   key: string,
   message: string
